+++ src/melange_json.mli
@@ -416,24 +416,10 @@ val parse : string -> json option [@@deprecated "Use `of_string` instead"]
 val parseOrRaise : string -> json [@@deprecated "Use `of_string` instead"]
 val stringify : json -> string [@@deprecated "Use `to_string` instead"]
 
-val classify :
-  json ->
-  [ `Assoc of (string * json) list
-  | `Bool of bool
-  | `Float of float
-  | `Int of int
-  | `List of json list
-  | `Null
-  | `String of string ]
-(** Classify a JSON value into a variant type. *)
+val classify : json -> Classify.t
+(** Classify a JSON value into a variant type, recursively. *)
 
 val declassify :
-  [ `Assoc of (string * json) list
-  | `Bool of bool
-  | `Float of float
-  | `Int of int
-  | `List of json list
-  | `Null
-  | `String of string ] ->
+  Classify.t ->
   json
 (** Declassify a variant type into a JSON value. *)